+++ ShowMeYourCloset/src/main/java/kh/spring/project/ClosetController.java
@@ -1,0 +1,5 @@
+package kh.spring.project;
+
+public class ClosetController {
+
+}
